+++ app/Http/Controllers/UsersController.php
@@ -11,14 +11,14 @@ class UsersController extends Controller
 {
     public function index(UsersDataTable $dt)
     {
-        user_can('manage-users');
+        $this->authorize('update', User::class);
 
         return $dt->render('user.index');
     }
 
     public function create()
     {
-        user_can('manage-users');
+        $this->authorize('update', User::class);
 
         $roles = Role::pluck('name', 'name');
 
@@ -27,7 +27,7 @@ class UsersController extends Controller
 
     public function store()
     {
-        user_can('manage-users');
+        $this->authorize('update', User::class);
 
         $this->validate(request(), [
             'name'                  => 'required|string',
@@ -56,7 +56,7 @@ class UsersController extends Controller
     */
     public function edit($id)
     {
-        user_can('manage-users');
+        $this->authorize('update', User::class);
 
         $user  = User::findOrFail($id);
         $roles = Role::pluck('name', 'name');
@@ -80,7 +80,7 @@ class UsersController extends Controller
     {
         $user = User::findOrFail($id) ?? auth()->user();
         if ($user != auth()->user()) {
-            user_can('manage-users');
+            $this->authorize('update', User::class);
         }
 
         $this->validate(request(), [
@@ -122,7 +122,7 @@ class UsersController extends Controller
     */
     public function destroy(User $user)
     {
-        user_can('manage-users');
+        $this->authorize('update', User::class);
 
         $user->delete();
 
@@ -138,7 +138,7 @@ class UsersController extends Controller
     */
     public function restore($id)
     {
-        user_can('manage-users');
+        $this->authorize('update', User::class);
 
         $user = User::withTrashed()->findOrFail($id);
         $user->restore();
